Partially rewrite django-suspense to use StreamingResponse
The original implementation relies on a task pool on the same worker. This does not work well with multiple workers and in the cloud. Also each suspense tag was using a http request which could cause DOS.

This new implementation uses a technique similar to NextJS with a streaming response. It also requires to use an ASGI server. WSGI and runserver will unfortunately treat the suspense requests as synchronous.